Refactor LTC/NTP code: types, naming & safety
Modernize and clean up LTC and NTP code across multiple files.

- Update copyright years to 2026.
- Rework LtcGenerator API and internals: rename enums and enum values to k-prefixed identifiers, rename/normalize member names (e.g. skip_free_, pitch_ticker_), change several signed ints to unsigned where appropriate (time/seconds/frame related), introduce kStopped/kStarted/kLimit state constants, tighten function signatures (use uint32_t, nullptr), and improve callback/handler implementations. These changes improve naming consistency and type safety.
- Improve HandleRequest/Input, SetSkip/SetPitch, Action* methods and timecode handling with safer casts, nullptr usage and clearer control flow.
- SystimeReader: change time_previous_ to uint32_t, use gettimeofday(nullptr) and explicit casts for time calculations.
- NtpServer: update includes (network_config.h, network_udp.h), add explicit time_t/static_cast conversions for time calculations, compact StaticCallbackFunction, and minor formatting/structure fixes.
- ltc.cpp: formatting and style cleanups, static_casts, safer Itoa/itoa usages and parse_timecode improvements.

Overall these edits are aimed at code clarity, consistency, and reducing implicit conversions/warnings.

+++ lib-ltc/src/ltc.cpp
@@ -2,7 +2,7 @@
  * @file ltc.cpp
  *
  */
-/* Copyright (C) 2019-2025 by Arjan van Vught mailto:info@gd32-dmx.org
+/* Copyright (C) 2019-2026 by Arjan van Vught mailto:info@gd32-dmx.org
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -30,37 +30,36 @@
 
 #include "ltc.h"
 
-namespace ltc
-{
+namespace ltc {
 uint32_t g_nDisabledOutputs;
 ltc::Type g_Type;
 
-static constexpr char kTypes[5][ltc::timecode::TYPE_MAX_LENGTH + 1] = {"Film 24fps ", "EBU 25fps  ", "DF 29.97fps", "SMPTE 30fps", "----- -----"};
+static constexpr char kTypes[5][ltc::timecode::TYPE_MAX_LENGTH + 1] = {
+	"Film 24fps ", 
+	"EBU 25fps  ", 
+	"DF 29.97fps", 
+	"SMPTE 30fps", 
+	"----- -----"
+};
 
-const char* get_type(ltc::Type type)
-{
-    if (type > ltc::Type::UNKNOWN)
-    {
+const char* get_type(ltc::Type type) {
+    if (type > ltc::Type::UNKNOWN) {
         return kTypes[static_cast<uint32_t>(ltc::Type::UNKNOWN)];
     }
 
     return kTypes[static_cast<uint32_t>(type)];
 }
 
-const char* get_type()
-{
-    if (ltc::g_Type > ltc::Type::UNKNOWN)
-    {
+const char* get_type() {
+    if (ltc::g_Type > ltc::Type::UNKNOWN) {
         return kTypes[static_cast<uint32_t>(ltc::Type::UNKNOWN)];
     }
 
     return kTypes[static_cast<uint32_t>(ltc::g_Type)];
 }
 
-ltc::Type get_type(uint8_t fps)
-{
-    switch (fps)
-    {
+ltc::Type get_type(uint8_t fps) {
+    switch (fps) {
         case 24:
             return ltc::Type::FILM;
             break;
@@ -80,13 +79,11 @@ ltc::Type get_type(uint8_t fps)
     return ltc::Type::UNKNOWN;
 }
 
-void set_type(uint8_t fps)
-{
+void set_type(uint8_t fps) {
     ltc::g_Type = get_type(fps);
 }
 
-void init_timecode(char* timecode)
-{
+void init_timecode(char* timecode) {
     assert(timecode != nullptr);
 
     memset(timecode, ' ', ltc::timecode::CODE_MAX_LENGTH);
@@ -96,8 +93,7 @@ void init_timecode(char* timecode)
     timecode[ltc::timecode::index::COLON_3] = ':';
 }
 
-void init_systemtime(char* systemtime)
-{
+void init_systemtime(char* systemtime) {
     assert(systemtime != nullptr);
 
     memset(systemtime, ' ', ltc::timecode::SYSTIME_MAX_LENGTH);
@@ -106,12 +102,10 @@ void init_systemtime(char* systemtime)
     systemtime[ltc::systemtime::index::COLON_2] = ':';
 }
 
-static void Itoa(uint32_t value, char* buffer)
-{
+static void Itoa(uint32_t value, char* buffer) {
     auto* p = buffer;
 
-    if (value == 0)
-    {
+    if (value == 0) {
         *p++ = '0';
         *p = '0';
         return;
@@ -121,8 +115,7 @@ static void Itoa(uint32_t value, char* buffer)
     *p = static_cast<char>('0' + (value % 10U));
 }
 
-void itoa_base10(const struct ltc::TimeCode* ltc_timecode, char* timecode)
-{
+void itoa_base10(const struct ltc::TimeCode* ltc_timecode, char* timecode) {
     assert(ltc_timecode != nullptr);
     assert(timecode != nullptr);
 
@@ -132,27 +125,23 @@ void itoa_base10(const struct ltc::TimeCode* ltc_timecode, char* timecode)
     Itoa(ltc_timecode->frames, &timecode[ltc::timecode::index::FRAMES]);
 }
 
-void itoa_base10(const struct tm* local_time, char* system_time)
-{
+void itoa_base10(const struct tm* local_time, char* system_time) {
     assert(local_time != nullptr);
     assert(system_time != nullptr);
 
-    Itoa(local_time->tm_hour, &system_time[ltc::systemtime::index::HOURS]);
-    Itoa(local_time->tm_min, &system_time[ltc::systemtime::index::MINUTES]);
-    Itoa(local_time->tm_sec, &system_time[ltc::systemtime::index::SECONDS]);
+    Itoa(static_cast<uint32_t>(local_time->tm_hour), &system_time[ltc::systemtime::index::HOURS]);
+    Itoa(static_cast<uint32_t>(local_time->tm_min), &system_time[ltc::systemtime::index::MINUTES]);
+    Itoa(static_cast<uint32_t>(local_time->tm_sec), &system_time[ltc::systemtime::index::SECONDS]);
 }
 
 #define DIGIT(x) ((x) - '0')
 #define VALUE(x, y) static_cast<uint8_t>(((x) * 10) + (y))
 
-bool parse_timecode(const char* time_code, uint8_t fps, struct ltc::TimeCode* timecode)
-{
+bool parse_timecode(const char* time_code, uint8_t fps, struct ltc::TimeCode* timecode) {
     assert(time_code != nullptr);
     assert(timecode != nullptr);
 
-    if ((time_code[ltc::timecode::index::COLON_1] != ':') || (time_code[ltc::timecode::index::COLON_2] != ':') ||
-        (time_code[ltc::timecode::index::COLON_3] != ':'))
-    {
+    if ((time_code[ltc::timecode::index::COLON_1] != ':') || (time_code[ltc::timecode::index::COLON_2] != ':') || (time_code[ltc::timecode::index::COLON_3] != ':')) {
         return false;
     }
 
@@ -160,22 +149,19 @@ bool parse_timecode(const char* time_code, uint8_t fps, struct ltc::TimeCode* ti
 
     auto tenths = DIGIT(time_code[ltc::timecode::index::FRAMES_TENS]);
 
-    if ((tenths < 0) || (tenths > 3))
-    {
+    if ((tenths < 0) || (tenths > 3)) {
         return false;
     }
 
     auto digit = DIGIT(time_code[ltc::timecode::index::FRAMES_UNITS]);
 
-    if ((digit < 0) || (digit > 9))
-    {
+    if ((digit < 0) || (digit > 9)) {
         return false;
     }
 
     auto value = VALUE(tenths, digit);
 
-    if (value >= fps)
-    {
+    if (value >= fps) {
         return false;
     }
 
@@ -185,21 +171,18 @@ bool parse_timecode(const char* time_code, uint8_t fps, struct ltc::TimeCode* ti
 
     tenths = DIGIT(time_code[ltc::timecode::index::SECONDS_TENS]);
 
-    if ((tenths < 0) || (tenths > 5))
-    {
+    if ((tenths < 0) || (tenths > 5)) {
         return false;
     }
 
     digit = DIGIT(time_code[ltc::timecode::index::SECONDS_UNITS]);
-    if ((digit < 0) || (digit > 9))
-    {
+    if ((digit < 0) || (digit > 9)) {
         return false;
     }
 
     value = VALUE(tenths, digit);
 
-    if (value > 59)
-    {
+    if (value > 59) {
         return false;
     }
 
@@ -209,22 +192,19 @@ bool parse_timecode(const char* time_code, uint8_t fps, struct ltc::TimeCode* ti
 
     tenths = DIGIT(time_code[ltc::timecode::index::MINUTES_TENS]);
 
-    if ((tenths < 0) || (tenths > 5))
-    {
+    if ((tenths < 0) || (tenths > 5)) {
         return false;
     }
 
     digit = DIGIT(time_code[ltc::timecode::index::MINUTES_UNITS]);
 
-    if ((digit < 0) || (digit > 9))
-    {
+    if ((digit < 0) || (digit > 9)) {
         return false;
     }
 
     value = VALUE(tenths, digit);
 
-    if (value > 59)
-    {
+    if (value > 59) {
         return false;
     }
 
@@ -234,22 +214,19 @@ bool parse_timecode(const char* time_code, uint8_t fps, struct ltc::TimeCode* ti
 
     tenths = DIGIT(time_code[ltc::timecode::index::HOURS_TENS]);
 
-    if ((tenths < 0) || (tenths > 2))
-    {
+    if ((tenths < 0) || (tenths > 2)) {
         return false;
     }
 
     digit = DIGIT(time_code[ltc::timecode::index::HOURS_UNITS]);
 
-    if ((digit < 0) || (digit > 9))
-    {
+    if ((digit < 0) || (digit > 9)) {
         return false;
     }
 
     value = VALUE(tenths, digit);
 
-    if (value > 23)
-    {
+    if (value > 23) {
         return false;
     }
 
@@ -258,28 +235,24 @@ bool parse_timecode(const char* time_code, uint8_t fps, struct ltc::TimeCode* ti
     return true;
 }
 
-bool parse_timecode_rate(const char* timecode_rate, uint8_t& fps)
-{
+bool parse_timecode_rate(const char* timecode_rate, uint8_t& fps) {
     assert(timecode_rate != nullptr);
 
     const auto kTenths = DIGIT(timecode_rate[0]);
 
-    if ((kTenths < 0) || (kTenths > 3))
-    {
+    if ((kTenths < 0) || (kTenths > 3)) {
         return false;
     }
 
     const auto kDigit = DIGIT(timecode_rate[1]);
 
-    if ((kDigit < 0) || (kDigit > 9))
-    {
+    if ((kDigit < 0) || (kDigit > 9)) {
         return false;
     }
 
     const auto kValue = VALUE(kTenths, kDigit);
 
-    switch (kValue)
-    {
+    switch (kValue) {
         case 24:
             fps = 24;
             ltc::g_Type = ltc::Type::FILM;
